7/2/22

+++ zee5appspringboot/src/main/java/com/zee/zee5app/repository/UserRepository.java
@@ -1,6 +1,7 @@
 package com.zee.zee5app.repository;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -14,6 +15,6 @@ public interface UserRepository extends JpaRepository<Register, String> {
 	//its pre-defined in jpa
 	//Boolean existsByEmail(String email);
 	//Boolean existsByContactNumber(BigDecimal contactNumber);
-	Boolean existsByEmailAndContactNumber(String email,BigDecimal contactNumber);
+	Boolean existsByEmailAndContactNumber(String email,BigInteger contactNumber);
 	
 }
